Fix anemic cache partition on SGS (5/6)
The cache partition is 80 MB, which should be enough to handle
almost anything we throw at it, except we also place our system's
dalvik-cache there, causing almost half of it to be wasted. This
causes the market to assume the device is running out of space
and is incapable of downloading certain applications (>40 MB).
Moving the cache to /data/cache doesn't seem to have any noticeable
speed decrease but allows us to actually download our apps.

Change-Id: I16a99639f652cc3ce5731fff201055ac12b68721